% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_lai}
\alias{calc_lai}
\title{Function that returns LAI values at the plot scale
#'}
\usage{
calc_lai()
}
\value{
A data frame of leaf area fro SLA * mass at the plot scale
}
\description{
Function that returns LAI values at the plot scale
#'
}
\examples{
fortedata::calc_lai()
}
