% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canopy_structural_traits}
\alias{canopy_structural_traits}
\title{Canopy Structural Traits from 2D Canopy LiDAR}
\format{A data frame with lots of variables:  \code{subplotID}, \code{year}, \code{mean.height},
\code{height.2}, \code{mean.height.var}, \code{mean.height.rms}, \code{transect.length}, \code{mode.el},
\code{max.el}, \code{mode.2}, \code{max.can.ht}, \code{mean.max.ht}, \code{mean.vai},
\code{mean.peak.vai}, \code{deep.gaps}, \code{deep.gap.fraction}, \code{porosity},
\code{std.std}, \code{mean.std}, \code{rugosity}, \code{top.rugosity}, \code{mean.return.ht},
\code{sd.return.ht}, \code{sky.fraction}, \code{cover.fraction}, \code{max.ht}, \code{scan.density},
\code{rumple}, \code{clumping.index}, and \code{enl}

\describe{
\item{subplotID}{subplot identifier}
\item{year}{msmt year}
\item{mean.height}{mean height of vai}
\item{height.2}{standard deviation of mean height}
\item{mean.height.var}{variance of mean height}
\item{mean.height.rms}{variance of mean height}
\item{transect.length}{length of transect}
\item{mode.el}{i forgot what this is}
\item{max.el}{greatest density of VAI x, z position}
\item{mode.2}{variance of maximum VAI}
\item{max.can.ht}{maximum measured canopy height}
\item{mean.max.ht}{mean outer canopy height or MOCH}
\item{mean.vai}{average VAI across transect}
\item{mean.peak.vai}{average height of maximum VAI}
\item{deep.gaps}{number of 1 m wide bins with no lidar returns}
\item{deep.gap.fraction}{deep gaps dividied by transect length}
\item{porosity}{ratio of empty to filled bins in the canopy}
\item{std.std}{precursor to rugosity}
\item{mean.std}{precursor to rugosity}
\item{rugosity}{accumulated canopy complexity metric}
\item{top.rugosity}{standard deviation of final lidar returns}
\item{mean.return.ht}{average lidar return distance}
\item{sd.return.ht}{standard deviation of lidar return distances}
\item{sky.fraction}{ratio of sky hits to lidar returns}
\item{cover.fraction}{1/sky fraction}
\item{max.ht}{same as max can ht, removed in later forestr updates}
\item{scan.density}{no. of lidar returns divided by transect length}
\item{rumple}{outer surface variability divided by transect lenght}
\item{clumping.index}{clumpiness}
\item{enl}{effective number of layers}
}}
\usage{
data(canopy_structural_traits)
}
\description{
Canopy strucutral Traits derived using the forestr 1.0.1 package from
2D portable canopy lidar
}
\keyword{datasets}
