% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_remote_sensing.R
\name{fd_canopy_structure_summary}
\alias{fd_canopy_structure_summary}
\title{Summary data for canopy structural data including canopy complexity and leaf area
by replicate by year for 2018 to 2020}
\usage{
fd_canopy_structure_summary()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Summary data for canopy structural data including canopy complexity and leaf area
by replicate by year for 2018 to 2020
}
\details{
The columns are as follows:
\itemize{
\item \code{replicate} (character): Replicate code, extracted from \code{subplot_id}.
\item \code{plot} (integer): Plot ID number, extracted from \code{subplot_id}.
\item \code{subplot} (character): Subplot code, extracted from \code{subplot_id}.
\item \code{year} (integer): Year in which measurement was taken
\item \code{rugosity} (numeric): mean of leaf area index
\item \code{rugosity_sd} (numeric): sd of leaf area index
\item \code{rugosity_n} (integer): number of observations per sample
\item \code{rugosity_se} (numeric): se of leaf area index
\item \code{vai_mean} (numeric): mean of leaf area index
\item \code{vai_mean_sd} (numeric): sd of leaf area index
\item \code{vai_mean_n} (integer): number of observations per sample
\item \code{vai_mean_se} (numeric): se of leaf area index
}
}
\note{
This summary reports only canopy structural complexity (as canopy rugosity)
and leaf area (as VAI, or vegetation area index--akin to PAI or LAI).
}
\examples{
fd_canopy_structure_summary()
}
\author{
Measurements by Jeff Atkins at the University of Michigan Biological Station.
}
