% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_litter.R
\name{fd_cwd}
\alias{fd_cwd}
\title{Coarse wood debris survey data}
\usage{
fd_cwd()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
These data include (\code{fwd} column) in grams collected annually every ~November from the litter traps located in each FoRTE  subplot. The \code{fraction} column informs the user is each row is either leaves which are broken down by species in the \code{species} column; misc which are small, unidentifiable leaf fragements or organic material; or fwd, for fine woody debris which are small twigs and sticks. For 2018-2020, there are four litter traps in each subplot. Each littertrap is circular (1.82 m circumference, 0.29 m radius, 0.26 m^2 area).See \code{fd_experimental_design_vignette} for more information. Within the dataset, the \code{bagtare_g} column is the mass of the measuring bag or tare and \code{bagmass_g} is the total mass of bag + sample.
}
\note{
Data were collected by multiple Gough Lab team members
}
\examples{
fd_litter()
}
