% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_inventory.R
\name{fd_dendro}
\alias{fd_dendro}
\title{Dendroband readings for canopy trees}
\usage{
fd_dendro()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
A data set including measurements taken from dendrobands, which give fine scale readings with a precision of 0.01 inch, allowing for
quantifying small, incremental growth not possible with DBH tapes. These data were collected using custom made steel band dendrometer bands fixed to a subsample of trees at ~1.3 height. Custom ruler stickers were used to measure the incremental circumference changes in inches. Dendrometer bands were fixed to ~ 700 trees in the summer of 2018. Measurements began in November 2018 with weekly measurements over the 2019 growing season and annual or bi-annual measurements in subsequent years.
}
\note{
Data were collected by multiple Gough Lab team members
}
\examples{
fd_dendro()
}
\author{
Maxim S Grigri
}
