% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_remote_sensing.R
\name{fd_hemi_camera_summary}
\alias{fd_hemi_camera_summary}
\title{Return summary data for hemispherical camera data}
\usage{
fd_hemi_camera_summary()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Return summary data for hemispherical camera data
}
\details{
The columns are as follows:
\itemize{
\item \code{Replicate} (character): Replicate code, extracted from \code{SubplotID}.
\item \code{Plot} (integer): Plot ID number, extracted from \code{SubplotID}.
\item \code{Subplot} (character): Subplot code, extracted from \code{SubplotID}.
\item \code{Date} (date): Date of measurement
\item \code{Year} (integer): Year of
\item \code{LAI_cam} (numeric): mean of leaf area index
\item \code{LAI_cam_sd} (numeric): sd of leaf area index
\item \code{LAI_cam_n} (integer): number of observations per sample
\item \code{LAI_cam_se} (numeric): se of leaf area index
}
}
\note{
For now this is pretty basic. More detailed summaries could be made,
e.g. by live/dead, species, etc.
}
\examples{
fd_hemi_camera_summary()
}
