% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_inventory.R
\name{fd_inventory_summary}
\alias{fd_inventory_summary}
\title{Basic statistics generated from the raw inventory data.}
\usage{
fd_inventory_summary()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Basic statistics generated from the raw inventory data.
}
\details{
The returned columns are as follows:
\itemize{
\item \code{Replicate} (character): Replicate code, extracted from \code{SubplotID}.
\item \code{Plot} (integer): Plot ID number, extracted from \code{SubplotID}.
\item \code{Subplot} (character): Subplot code, extracted from \code{SubplotID}.
\item \code{BA_m2_ha} (numeric): Basal area, square meters per hectare.
\item \code{Stocking_ha} (numeric): Stocking, trees per hectare.
}
}
\note{
For now this is pretty basic. More detailed summaries could be made,
e.g. by live/dead, species, etc.
}
\examples{
fd_inventory_summary()
}
