% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_litter.R
\name{fd_lai}
\alias{fd_lai}
\title{Raw inventory table.}
\usage{
fd_lai()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Raw inventory table.
}
\details{
The columns are as follows:

SubplotID	Species	bag_no	bag_tare_g	bag_mass_g
\itemize{
\item \code{SubplotID} (character): Subplot ID number. These subplot codesarea
concatenation of the plot (\code{\link{fd_plots}}) and
subplot \code{\link{fd_subplots}} codes.
#' - \code{Replicate} (character): Replicate code, extracted from \code{SubplotID}.
\item \code{Plot} (integer): Plot ID number, extracted from \code{SubplotID}.
\item \code{Year} (integer): year of litter collection
\item \code{Subplot} (character): Subplot code, extracted from \code{SubplotID}.
\item \code{Species} (character): Species code from the USDA Plants Database; see
\url{https://plants.sc.egov.usda.gov/java/}.
\item \code{BagTare_g} (numeric): Diameter at breast height (1.37m), cm.
\item \code{BagMass_g} (numeric): weight of bag + litter
}
}
\examples{
fd_lai()
}
