% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_litter.R
\name{fd_litter}
\alias{fd_litter}
\title{Litter trap data.}
\usage{
fd_litter()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Litter trap data.
}
\examples{
fd_litter()
}
\author{
Measurements by Gough Lab at the University of Michigan Biological Station.
}
