% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_make_release.R
\name{fd_make_release}
\alias{fd_make_release}
\title{Write files for a GitHub data release}
\usage{
fd_make_release(path, zip_release = TRUE)
}
\arguments{
\item{path}{Path to write files to, character}

\item{zip_release}{Zip files into a single archive? Logical}
}
\value{
Fully-qualified filename of release file
}
\description{
Write files for a GitHub data release
}
\note{
This function creates output that is mirrored at https://figshare.com/articles/dataset/fortedata-1_0_1/12292490
}
