% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fd_metadata}
\alias{fd_metadata}
\title{Data tables' metadata.}
\usage{
fd_metadata(table = NULL)
}
\arguments{
\item{table}{Name of table to return, character, optional}
}
\value{
A \code{data.frame} or \code{tibble} holding field metadata.
}
\description{
Data tables' metadata.
}
