% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_belowground.R
\name{fd_micromet}
\alias{fd_micromet}
\title{Return belowground micrometeorology data}
\usage{
fd_micromet()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Return belowground micrometeorology data
}
\details{
The columns are as follows:
\itemize{
\item \code{SubplotID} (character): Subplot ID number. These subplot codes are a
concatenation of the plot (\code{\link{fd_plots}}) and
subplot \code{\link{fd_subplots}} codes.
\item \code{Replicate} (character): Replicate code, extracted from \code{SubplotID}.
\item \code{Plot} (integer): Plot ID number, extracted from \code{SubplotID}.
\item \code{Subplot} (character): Subplot code, extracted from \code{SubplotID}.
\item \code{Year} (integer): Year of mesmt
\item \code{Date} (character) Date of measurement in YYYY-MM-DD
\item \code{DateTime} (POSIXlt) Format in "\%Y-\%m-\%d \%H:\%M:\%S"
\item \code{NestedPlot} (integer): Nested subplot sampling point inside subplot
\item \code{soilTemp} (numeric): soil temperature measured to 7 cm depth in celsius
\item \code{VWC} (numeric): volumetric water content in percent
}
}
\examples{
fd_micromet()
}
