% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_belowground.R
\name{fd_micromet_summary}
\alias{fd_micromet_summary}
\title{Return basic statistics generated from soil micrometeorology}
\usage{
fd_micromet_summary()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Return basic statistics generated from soil micrometeorology
}
\details{
The returned columns are as follows:
\itemize{
\item \code{Replicate} (character): Replicate code, extracted from \code{SubplotID}.
\item \code{Plot} (integer): Plot ID number, extracted from \code{SubplotID}.
\item \code{Subplot} (character): Subplot code, extracted from \code{SubplotID}.
\item \code{Date} (date): Date of measurements.
\item \code{soilTemp} (numeric): Mean of soil temperature
\item \code{soilTemp_sd} (numeric): Standard Deviation of soil temperature.
\item \code{soilTemp_n} (integer): number of soil temperature msmts per
\item \code{soilTemp_se} (numeric): Standard error of soil temperature.
\item \code{VWC} (numeric): Mean of volumetric water content.
\item \code{VWC_sd} (numeric): Standard Deviation of volumetric water content.
\item \code{VWC_n} (integer): Number of VWC msmts per
\item \code{VWC_se} (numeric): Standard error of volumetric water content
}
}
\note{
For now this is pretty basic. More detailed summaries could be made
}
\examples{
fd_micromet_summary()
}
