% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_inventory.R
\name{fd_mortality}
\alias{fd_mortality}
\title{Mortality Assignments for 2018 Stem Girdling}
\usage{
fd_mortality()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.#'
}
\description{
A data set similar to \code{fd_inventory} but includes the column \code{fate} which has a value of either "kill" or "live" with "kill" indicating the tree was stem girdled.
}
\examples{
fd_mortality()
}
\author{
Atkins
}
