% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_plots.R
\name{fd_nested_subplots}
\alias{fd_nested_subplots}
\title{Nested subplots table.}
\usage{
fd_nested_subplots()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Nested subplots table.
}
\examples{
fd_nested_subplots()
}
\author{
Measurements by Gough Lab at the University of Michigan Biological Station.
}
