% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_observations.R
\name{fd_observations}
\alias{fd_observations}
\title{Number of observations in the FoRTE Dataset}
\usage{
fd_observations()
}
\description{
Number of observations in the FoRTE Dataset
}
\note{
No. of observations for many, but not all datasets

#' @return A \code{data.frame} or \code{tibble} of the observation tally.
Call \code{\link{fd_metadata}} for field metadata.
}
\examples{
fd_observations()
}
