% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_remote_sensing.R
\name{fd_par_summary}
\alias{fd_par_summary}
\title{Summary data for ceptometer.}
\usage{
fd_par_summary()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Summary data for ceptometer.
}
\details{
The columns are as follows:
\itemize{
\item \code{Replicate} (character): Replicate code, extracted from \code{SubplotID}.
\item \code{Plot} (integer): Plot ID number, extracted from \code{SubplotID}.
\item \code{Subplot} (character): Subplot code, extracted from \code{SubplotID}.
\item \code{Date} (date): Date of measurement
\item \code{faPAR} (numeric): mean of the fraction of absorbed PAR
\item \code{faPAR_sd} (numeric): sd of the fraction of absorbed PAR
\item \code{faPAR_n} (integer): number of observations per sample
\item \code{faPAR_se} (numeric): se of the fraction of absorbed PAR
\item \code{LAI_cept} (numeric): mean of leaf area index
\item \code{LAI_cept_sd} (numeric): sd of leaf area index
\item \code{LAI_cept_n} (integer): number of observations per sample
\item \code{LAI_cept_se} (numeric): se of leaf area index
}
}
\note{
For now this is pretty basic.
}
\examples{
fd_par_summary()
}
