% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fd_plot_metadata}
\alias{fd_plot_metadata}
\title{Function that brings in all plot metadata for assignment}
\usage{
fd_plot_metadata()
}
\value{
gives the plot info
}
\description{
Function that brings in all plot metadata for assignment
}
\examples{
fortedata::fd_plot_metadata()
}
