% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_plots.R
\name{fd_plots}
\alias{fd_plots}
\title{Return the plots table.}
\usage{
fd_plots()
}
\value{
A `data.frame`. See "Details" for column descriptions.
}
\description{
Return the plots table.
}
\details{
The columns are as follows:
- `Replicate` (character): Replicate code. Each replicate contains
  multiple plots with different disturbance treatments.
- `Plot` (integer): Plot ID number. Plots are nested within replicates.
- `Latitude`, `Longitude` (double): Plot coordinates, in decimal
  degrees.
- `Disturbance_severity` (double): Level of disturbance severity,
  expressed as a percent (0 - 100) reduction in the target variable (TBD).
}
\examples{
fd_plots()
}
