% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_publications.R
\name{fd_publications}
\alias{fd_publications}
\title{FoRTE related publications}
\format{
A data frame with
\describe{
\item{entry}{Abbreviated in-text entry in the form of author last name and year}
\item{pub_year}{integer year of publication}
\item{title}{character, publication title}
\item{journal}{character, publishing journal}
\item{doi}{character, the Digital Object Identifier}
}
}
\usage{
fd_publications()
}
\description{
Table of publications produced as part of the NSF Forest Resilience Threshold Experiment
Project.
}
\examples{
\dontrun{
fd_publications()
}
}
