% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_inventory.R
\name{fd_seedling_sapling}
\alias{fd_seedling_sapling}
\title{Seedling and sapling data}
\usage{
fd_seedling_sapling()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Seedling and sapling data
}
\note{
Data were collected by Maxim S. Grigri
}
\examples{
fd_seedling_sapling()
}
