% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_belowground.R
\name{fd_soil_respiration}
\alias{fd_soil_respiration}
\title{Soil respiration (soil to atmosphere CO2 efflux) table.}
\usage{
fd_soil_respiration()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Soil respiration (soil to atmosphere CO2 efflux) table.
}
\note{
Measurements taken by Kayla Mathes and Carly Rodriguez
}
\examples{
fd_soil_respiration()
}
