% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_belowground.R
\name{fd_soilCO2_summary}
\alias{fd_soilCO2_summary}
\title{Return basic statistics generated from soil co2 effluxx.}
\usage{
fd_soilCO2_summary()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Return basic statistics generated from soil co2 effluxx.
}
\details{
The returned columns are as follows:
\itemize{
\item \code{Replicate} (character): Replicate code, extracted from \code{SubplotID}.
\item \code{Plot} (integer): Plot ID number, extracted from \code{SubplotID}.
\item \code{Subplot} (character): Subplot code, extracted from \code{SubplotID}.
\item \code{Date} (date): Date of measurements.
\item \code{soilCO2Efflux} (numeric): Mean of Soil Co2 Efflux.
\item \code{soilCO2Efflux_sd} (numeric): Standard Deviation of Soil CO2 Efflux.
\item \code{soilCO2Efflux_n} (integer): number of soil CO2 Efflux msmts per
\item \code{soilCO2Efflux_se} (numeric): Standard error of Soil Co2 Efflux.
}
}
\note{
For now this is pretty basic. More detailed summaries could be made,
e.g. by live/dead, species, etc.
}
\examples{
fd_soilCO2_summary()
}
