% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_subcanopy_density.R
\name{fd_subcanopy_density}
\alias{fd_subcanopy_density}
\title{Subcanopy density data}
\usage{
fd_subcanopy_density()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Subcanopy density data
}
\note{
Data were collected by multiple Gough Lab team members
}
\examples{
fd_subcanopy_density()
}
