% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_subcanopy_diameter.R
\name{fd_subcanopy_diameter}
\alias{fd_subcanopy_diameter}
\title{Subcanopy diameter data}
\usage{
fd_subcanopy_diameter()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Subcanopy diameter data
}
\note{
Data were collected by multiple Gough Lab team members
}
\examples{
fd_subcanopy_diameter()
}
