% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_plots.R
\name{fd_treatments}
\alias{fd_treatments}
\title{Treatment assignments}
\usage{
fd_treatments()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Treatment assignments
}
\details{
The columns are as follows:
\itemize{
\item \code{replicate} (character): Replicate code. Each replicate contains
multiple plots with different disturbance treatments.#'
\item \code{plot} (integer): plot code
one in the eastern part, one in the western.
\item \code{subplot} (double): Area of the nested subplot, in square meters.
\item \code{disturbance_severity} (integer): assigned targeted distrubance severity
percentage, based on leaf area
\item \code{treatment} (character): either top as T or bottomup B
}
}
\examples{
fd_treatments()
}
\author{
Measurements by Gough Lab at the University of Michigan Biological Station.
}
