% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leaf_spectrometry}
\alias{leaf_spectrometry}
\title{Leaf-level spectral indices.}
\format{A data frame with four variables: \code{subplotID}, \code{species},
\code{index}, and \code{value}

\describe{
\item{subplotID}{subplot identifier}
\item{species}{species of leaf analyzed}
\item{index}{spectral index}
\item{value}{value corresponding to the spectral index}
}
@source \url{http://atkinsjeff.github.io}}
\usage{
data(leaf_spectrometry)
}
\description{
Canopy leaf spectra collected with a CI-710
Miniature Leaf SpectrometerARI1	 #FF0000	(1/W550) - (1/W700)	 Leaf Pigments	 The value of this index ranges from 0 to more than 0.2. The common range for green vegetation is 0.001 to 0.1.
subplotID	species	index	value
}
\keyword{datasets}
