% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_metadata}
\alias{plot_metadata}
\title{Function that brings in all plot metadata for assignment}
\usage{
plot_metadata()
}
\value{
gives the plot info
}
\description{
Function that brings in all plot metadata for assignment
}
\examples{
fortedata::plot_metadata()
}
