% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_csv_file}
\alias{read_csv_file}
\title{Read an internal package CSV file.}
\usage{
read_csv_file(...)
}
\arguments{
\item{...}{Filename, character}
}
\value{
The read-in dataframe.
}
\description{
Read an internal package CSV file.
}
\details{
\code{...} is passed to \code{\link{system.file}},
so can be vectors specifying
subdirectory and file(s) within the package.
}
\note{
This is an internal function and not intended to be called directly.
}
\keyword{internal}
