% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_efflux}
\alias{soil_efflux}
\title{Soil CO2 Efflux measuremnts and Associated Micrometeorological Measurements}
\format{A data frame with four variables: \code{dateTime} \code{subplotID}, \code{run},\code{nestedPlot},
\code{soilCO2Efflux}, \code{soilTemp}, \code{VWC} and \code{notes}

\describe{
\item{dateTime}{date and time of sample in YYYY-MM-DD HH:MM:SS}
\item{subplotID}{subplot identifier}
\item{run}{differnt sampling occurence}
\item{nestedPlot}{sampling point of collar inside of subplot}
\item{soilCO2Efflux}{soil carbon dioxide efflux in umol CO2 m^2 s^-1}
\item{soilTemp}{soil temperature at 7 cm depth in celsius}
\item{VWC}{volumetric water content from 0 to 20 cm depth}
\item{notes}{additional information}
}
@source \url{http://atkinsjeff.github.io}}
\usage{
data(soil_efflux)
}
\description{
Soil CO efflux measurements collected with a Li-Cor 8100A
Soil temperature measurements were made at 7 cm depth
soil moisture as volumetric water content msmts were wmade with a Campbell CS620
time domain reflectometer (TDR) at 0 - 20 cm depth.
dateTime	subplotID	nestedPlot	run	soilCO2Efflux	soilTemp	VWC	notes
}
\keyword{datasets}
