% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_subplot_id}
\alias{split_subplot_id}
\title{Split the SubplotID column into more useful individual columns}
\usage{
split_subplot_id(df)
}
\arguments{
\item{df}{A data.frame with a four-character \code{subplot_id} column}
}
\value{
The data frame with new columns \code{replicate}, \code{plot},
and \code{subplot}.
}
\description{
Split the SubplotID column into more useful individual columns
}
\examples{
fortedata::split_subplot_id(data.frame(subplot_id = "A01E"))
}
\keyword{internal}
