% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_lai}
\alias{calc_lai}
\title{Function that returns LAI values at the plot scale}
\usage{
calc_lai()
}
\value{
A data frame of leaf area fro SLA * mass at the plot scale
}
\description{
LAI is calculated using the equation SLA * mass, where SLA is specific leaf area. Species and site specific values
are in fd_sla.csv
}
\examples{
fortedata::calc_lai()
}
