% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_remote_sensing.R
\name{fd_canopy_structure}
\alias{fd_canopy_structure}
\title{Canopy structural traits from 2D canopy LiDAR.}
\usage{
fd_canopy_structure()
}
\value{
A \code{data.frame} or \code{tibble} of hemispherical camera data.
Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Canopy structural traits from 2D canopy LiDAR.
}
\note{
Data were collected by Jeff W. Atkins and Brandon Alveshare using an upward facing, portable canopy lidar unit.
The Canopy structural traits were derived using the
\code{forestr} 1.0.1 package from 2D portable canopy lidar
}
\examples{
fd_canopy_structure()
}
