% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_remote_sensing.R
\name{fd_canopy_structure_summary}
\alias{fd_canopy_structure_summary}
\title{Return summary data for hemispherical camera data}
\usage{
fd_canopy_structure_summary()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Return summary data for hemispherical camera data
}
\details{
The columns are as follows:
\itemize{
\item \code{Replicate} (character): Replicate code, extracted from \code{SubplotID}.
\item \code{Plot} (integer): Plot ID number, extracted from \code{SubplotID}.
\item \code{Subplot} (character): Subplot code, extracted from \code{SubplotID}.
\item \code{Date} (date): Date of measurement
\item \code{Year} (integer): Year of
\item \code{rugosity} (numeric): mean of leaf area index
\item \code{rugosity_sd} (numeric): sd of leaf area index
\item \code{rugosity_n} (integer): number of observations per sample
\item \code{rugosity_se} (numeric): se of leaf area index
\item \code{mean.vai} (numeric): mean of leaf area index
\item \code{mean.vai_sd} (numeric): sd of leaf area index
\item \code{mean.vai_n} (integer): number of observations per sample
\item \code{mean.vai_se} (numeric): se of leaf area index
}
}
\note{
For now this is pretty basic.
}
\examples{
fd_canopy_structure_summary()
}
