% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_remote_sensing.R
\name{fd_ceptometer}
\alias{fd_ceptometer}
\title{Ceptometer data.}
\usage{
fd_ceptometer()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Ceptometer data.
}
\examples{
fd_ceptometer()
}
\author{
Measurements by Jeff Atkins at the University of Michigan Biological Station.
}
