% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_inventory.R
\name{fd_dendro}
\alias{fd_dendro}
\title{Dendroband readings for canopy trees}
\usage{
fd_dendro()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.#'
}
\description{
A data set including measurements taken from dendrobands, which give fine scale readings with a precision of 0.01 inch, allowing for
quantifying small, incremental growth not possible with DBH tapes.
}
\examples{
fd_dendro()
}
\author{
Grigri
}
