% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_remote_sensing.R
\name{fd_hemi_camera}
\alias{fd_hemi_camera}
\title{Hemispherical camera data.}
\usage{
fd_hemi_camera()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Hemispherical camera data.
}
\examples{
fd_hemi_camera()
}
\author{
Measurements by Jeff Atkins at the University of Michigan Biological Station.
}
