% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_inventory.R
\name{fd_inventory}
\alias{fd_inventory}
\title{Forest inventory data for the 32 FoRTE subplots. Includes diameter-at-breast-height
(DBH) measurements in units of cm as well as information on health (i.e. see
\code{health_status} column where D = dead, M = moribund, and L = live) and canopy status
(i.e. see \code{canopy_status} where OD = overstory dominant, UN = understory,
OS = overstory submissive, SA = sapling,  and NA is a blank or missing record)}
\usage{
fd_inventory()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Forest inventory data for the 32 FoRTE subplots. Includes diameter-at-breast-height
(DBH) measurements in units of cm as well as information on health (i.e. see
\code{health_status} column where D = dead, M = moribund, and L = live) and canopy status
(i.e. see \code{canopy_status} where OD = overstory dominant, UN = understory,
OS = overstory submissive, SA = sapling,  and NA is a blank or missing record)
}
\note{
Data were collected by Gough Lab team members Autym Shafer,
Catherine McGuigan, and Alexandra Barry in 2018
using a Haglof Postex Inventory Unit with DPII Caliper
Where 'species' column is '????" indicates unknown plant id to be corrected
in future work.
}
\examples{
fd_inventory()
}
