% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_inventory.R
\name{fd_inventory}
\alias{fd_inventory}
\title{Raw inventory table.}
\usage{
fd_inventory()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.#'
}
\description{
Raw inventory table.
}
\examples{
fd_inventory()
}
\author{
Measurements by Gough Lab at the University of Michigan Biological Station.
}
