% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_inventory.R
\name{fd_inventory}
\alias{fd_inventory}
\title{Raw inventory table.}
\usage{
fd_inventory()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Raw inventory table.
}
\details{
The columns are as follows:
\itemize{
\item \code{Site} (character): Replicate code. Each replicate contains
multiple plots with different disturbance treatments.
\item \code{SubplotID} (character): Subplot ID number. These subplot codes are a
concatenation of the plot (\code{\link{fd_plots}}) and
subplot \code{\link{fd_subplots}} codes.
\item \code{Tag} (integer): Tree tag number.
\item \code{Species} (character): Species code from the USDA Plants Database; see
\url{https://plants.sc.egov.usda.gov/java/}.
\item \code{DBH_cm} (numeric): Diameter at breast height (1.37m), cm.
\item \code{Health_status} (character): live (L), moribund (M), or dead (D).
\item \code{Canopy_status} (character): Canopy status: overstory dominant (OD),
overstory submissive (OS), sapling (SA), understory (UN).
\item \code{Date} (date): Inventory entry date.
\item \code{Notes} (character): Notes.
\item \code{Replicate} (character): Replicate code, extracted from \code{SubplotID}.
\item \code{Plot} (integer): Plot ID number, extracted from \code{SubplotID}.
\item \code{Subplot} (character): Subplot code, extracted from \code{SubplotID}.
}
}
\examples{
fd_inventory()
}
\author{
Measurements by Gough Lab at the University of Michigan Biological Station.
}
