% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_leaf_physiology.R
\name{fd_leaf_spectrometry}
\alias{fd_leaf_spectrometry}
\title{Leaf spectrometry data}
\usage{
fd_leaf_spectrometry()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Leaf spectrometry data
}
\note{
Data were collected by Lisa Haber, Laura Hickey, Alexandra Barry, and Autym Shafer
}
\examples{
fd_leaf_spectrometry()
}
