% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_inventory.R
\name{fd_mortality}
\alias{fd_mortality}
\title{Mortality Assignments for Girdling 2018}
\usage{
fd_mortality()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.#'
}
\description{
Mortality Assignments for Girdling 2018
}
\examples{
fd_mortality()
}
\author{
Atkins
}
