% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_leaf_physiology.R
\name{fd_photosynthesis}
\alias{fd_photosynthesis}
\title{Leaf photosynthesis measurements.}
\usage{
fd_photosynthesis()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Leaf photosynthesis measurements.
}
\examples{
fd_photosynthesis()
}
\author{
Measurements by Lisa Haber at the University of Michigan Biological Station.
}
