% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_leaf_physiology.R
\name{fd_photosynthesis}
\alias{fd_photosynthesis}
\title{Leaf photosynthesis measurements.}
\usage{
fd_photosynthesis()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Leaf photosynthesis measurements.
}
\details{
The columns are as follows:
\itemize{
\item \code{Obs} (numeric) Observation number within file.
\item \code{HHMMSS} (character) Time of day.
\item \code{FTime} (integer) Number of seconds since logging began.
\item \code{EBal.} (integer) Energy balance on?
\item \code{Photo} (numeric) Photosynthetic rate in µmol CO2 m^-2 s^-1.
\item \code{Cond} (numeric) Stomatal conductance mol H2O m^-2 s^-1.
\item \code{Ci} (numeric) Intercellular CO2 concentration µmol CO2  mol^-1.
\item \code{Trmmol} (numeric) Transpiration rate mmol H2O m^-2 s^-1.
\item \code{VpdL} (numeric) Leaf-level vapor pressure deficit kPa.
\item \code{CTleaf} (numeric) Leaf temperature from energy balance calculation (degrees C).
\item \code{Area} (numeric) In-chamber leaf area (cm2; note that this area is not the true sampled tissue area for needleleaf samples).
\item \code{BLC_1} (numeric) One sided boundary layer conductance (mol/m2/s)
\item \code{StmRat} (integer) Stomatal ratio estimate
\item \code{BLCond}(numeric) Boundary layer conductance (mol/m2/s).
\item \code{Tair} (numeric) Chamber air temperature (degrees C).
\item \code{Tleaf} (numeric) Leaf surface temperature (degrees C).
\item \code{TBlk} (numeric) IRGA block temperature (degrees C)
\item \code{CO2R} (integer) Reference CO2 concentration (mu  mol/mol)
\item \code{CO2S} (integer) Sample CO2 concentration (µmol/mol)
\item \code{H2OR} (numeric) Reference H20 concentration (µmol/mol)
\item \code{H2OS} (numeric) Sample H20 concentration (µmol/mol)
\item \code{RH_R} (numeric) Reference relative humidity (\%)
\item \code{RH_S} (numeric) Sample relative humiditiy (\%)
\item \code{Flow}  (numeric) Flow rate (µmol/mol)
\item \code{PARi} (numeric) in-chamber PAR (µmol/m2/s)
\item \code{PARo} (numeric) external PAR (µmol/m2/s)
\item \code{Press} (numeric) Atmospheric pressure (kPa)
\item \code{CsMch} (numeric) standard deviation of CO2S during averaging time of most recent match
\item \code{HsMch} (numeric) standard deviation of H2OS during averaging time of most recent match
\item \code{CsMchSD} (numeric) standard deviation of CO2R during averaging time of most recent match
\item \code{HsMchSD} (numeric) standard deviation of H2OS during averaging time of most recent match
\item \code{CrMchSD} (numeric) standard deviation of CO2R during averaging time of most recent match
\item \code{HrMchSD} (numeric) standard deviation of H2OR during averaging time of most recent match
\item \code{StableF} (numeric) stable/total as a fraction
\item \code{BLCslope} (numeric) Slope term used in calculating boundary layer conductance
\item \code{BLCoffst} (numeric) Intercept term using in calculating boundary layer conductance
\item \code{f_parin} (integer) fraction of ParIn_um to use for energy balance
\item \code{f_parout} (integer) fraction of ParOut_um to use for energy balance
\item \code{alphaK} (numeric) used in conversion of µmol/mol to W/m2
\item \code{Status} (character) Status variable
\item \code{Plot} (integer): Plot ID number, extracted from \code{SubplotID}.
\item \code{Species} (character) Species code from the USDA Plants Database; see
\item \code{Sample} (character) Sample number--varies by species by plot
\item \code{Comments} (character) ID of closest vegetation survey plot (NE, SE, SW, NW) to
the stem measured, plus any additional comments
\item \code{Timestamp} (POSIXlt) Timestamp of measurement
\item \code{Replicate} (character): Replicate code
\item \code{Plot} (integer): Plot ID number
\item \code{Subplot} (character): Subplot code
}
}
\examples{
fd_photosynthesis()
}
\author{
Measurements by Lisa Haber at the University of Michigan Biological Station.
}
