% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_leaf_physiology.R
\name{fd_photosynthesis_summary}
\alias{fd_photosynthesis_summary}
\title{Return basic statistics generated from the raw licor data}
\usage{
fd_photosynthesis_summary()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Return basic statistics generated from the raw licor data
}
\details{
The returned columns are as follows:
\itemize{
\item \code{Replicate} (character): Replicate code, extracted from \code{SubplotID}.
\item \code{Plot} (integer): Plot ID number, extracted from \code{SubplotID}.
\item \code{Subplot} (character): Subplot code, extracted from \code{SubplotID}.
}
}
\note{
For now this one doesn't have everything
}
\examples{
fd_photosynthesis_summary()
}
