% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fd_plot_metadata}
\alias{fd_plot_metadata}
\title{Metadata for the experimental field plots}
\usage{
fd_plot_metadata()
}
\value{
A A \code{data.frame} or \code{tibble} with the experimental treatments by UMBS plot.
}
\description{
Metadata for the experimental field plots
}
\examples{
fd_plot_metadata()
}
