% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_plots.R
\name{fd_plots}
\alias{fd_plots}
\title{Plots table.}
\usage{
fd_plots()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Plots table.
}
\examples{
fd_plots()
}
\author{
Measurements by Gough Lab at the University of Michigan Biological Station.
}
