% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_plots.R
\name{fd_plots}
\alias{fd_plots}
\title{Plots table.}
\usage{
fd_plots()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Plots table.
}
\details{
The columns are as follows:
\itemize{
\item \code{Replicate} (character): Replicate code. Each replicate contains
multiple plots with different disturbance treatments.
\item \code{Plot} (integer): Plot ID number. Plots are nested within replicates.
\item \code{Longitude_plot}, \code{Latitude_plot} (double): Plot coordinates, in decimal
degrees.
\item \code{Plot_area_m2} (double): Area of the plot, in square meters.
}
}
\examples{
fd_plots()
}
