% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_belowground.R
\name{fd_soil_respiration}
\alias{fd_soil_respiration}
\title{Soil respiration (soil to atmosphere CO2 efflux) table}
\usage{
fd_soil_respiration()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
These data include soil CO2 efflux in units of micromoles per m^2 s^-1 as well as soil temperature in C, and soil water content as a percentage. See \code{fd_belowground_vignette} for further information.
}
\note{
Measurements taken by Kayla Mathes and Carly Rodriguez
}
\examples{
fd_soil_respiration()
}
