% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_plots.R
\name{fd_subplots}
\alias{fd_subplots}
\title{Subplots table.}
\usage{
fd_subplots()
}
\value{
A \code{data.frame} or \code{tibble}. See "Details" for column descriptions.
}
\description{
Subplots table.
}
\details{
The columns are as follows:
\itemize{
\item \code{Replicate} (character): Replicate code. Each replicate contains
multiple plots with different disturbance treatments.
\item \code{Plot} (integer): Plot ID number. Plots are nested within replicates.
\item \code{Subplot} (character): Subplot code. Subplots are nested within plots:
one in the eastern part, one in the western.
\item \code{Longitude_subplot}, \code{Latitude_subplot} (double): Subplot coordinates, in decimal
degrees.
\item \code{Plot_area_m2} (double): Area of the subplot, in square meters.
}
}
\examples{
fd_subplots()
}
