% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_plots.R
\name{fd_treatments}
\alias{fd_treatments}
\title{Treatment assignments}
\usage{
fd_treatments()
}
\value{
A \code{data.frame} or \code{tibble}. Call \code{\link{fd_metadata}} for field metadata.
}
\description{
Treatment assignments
}
\examples{
fd_treatments()
}
\author{
Measurements by Gough Lab at the University of Michigan Biological Station.
}
