% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forte_colors}
\alias{forte_colors}
\title{The FoRTE color palettte}
\usage{
forte_colors()
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes pulled from photographs of US National Parks.
}
\examples{
pal <- forte_colors()
}
\keyword{colors}
