% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forte_colors}
\alias{forte_colors}
\title{FoRTE color palette
The FoRTE color palette}
\usage{
forte_colors()
}
\value{
A vector of colors.
}
\description{
A FoRTE specific color palette used for color-classification of disturbance severity
treatments. 0\% = "#000000", 45\% = "#009E73", 65\% = "#0072B2", and 85\% = "#D55E00".
}
\details{
Per the example below, \code{forte_colors()} can be used to define a custom palette that
can then be used for plotting with other commands such as \code{scale_fill_manual()}, etc.
}
\examples{

 forte.pal <- forte_colors()

}
\keyword{colors}
\keyword{palette}
